[build] Use transition rule for multi-config build
Bazel has a "transition" rule that allows a target to be built using a
different set of configurations than the ones bazel is invoked with.
This makes it possible for a target to depend on other targets that need
to be built using a different setting, i.e. sysroot, arch-specific efi
images etc. The CL reworks build rule with transition rule such that it
is no longer necessary to pass "--platform" flags to select architecture
or build rust sysroot as a separate step.

Test: Build and run UEFI application/tests
Bug: 304282799

Change-Id: I38ef4fddac3693e70e9af7c6d257939056226f44