Simplify scratch/gpt buffer management
libstorage currently requires users to maintain a scratch and gpt
buffer and pass them everytime they use the APIs, which can be
inconvenient. This CL extends the BlockDevice trait with an interface
to allow users to provide an opaque scratch buffer in one place. The
APIs internally gets the buffer from the new interface, initializes and
partitions it for scratch and GPT and no longer requires them from the
arguments.

Change-Id: I10a2049c8159a6d92ff9bcd92cc03817f2253c81