Storage testlib can define TestBlockDevice from partitions
Add functionality to TestBlockDeviceBuilder to define a block device
as a collection of partitions with either associated data or just a size.
Partition sizes are rounded up to the block size and padded with
trailing zeroes.

If the device is built via partitions, a valid GPT header and backup
header are added to the device.

Tests: new unit test passes

Change-Id: I614f7d0c48e56b34ae9eaf08e5b0c953355f0cb0