slots::Manager.mark_boot_attempt takes no parameter
Remove the parameter from mark_boot_attempt.
If the return value of get_oneshot_status is a recovery target,
or the active slot is a recovery target,
does nothing but return the boot token.

If the current active slot is a normal slot,
decrements its tries and returns the boot token.

Tests: all unit tests pass
Change-Id: I7a8c84718c1bee32b9ad06acb892b9937f186cf9