Provide a slots::Manager over GBL AB slots protocol
Write an implementation for slots::Manager that uses the custom AB
slots EFI protocol on Android.

It is a vendor implementation detail whether or where this information
is backed on disk.

The wrapping allows higher level GBL code to abstract over whether the
platform is booting Android or Fuchsia and, if Android, vendor
implementation specifics.

Bug: b/345531636

Change-Id: Ie31a7e61b6734d7e33d08782b97d410a6f82b0bd