Change bootconfig VTS requirement
Don't require Android S launching devices to use bootconfig, even if the
functionality is available in some of the kernels that can be used.
Because we allowed Android S devices to launch with kernel version 4.19,
we need to continue supporting Android S devices that don't have
bootconfig support in the Android platform.
Android T is the first level that we only allow launching kernels 5.10+, so we
can expect all of them to use bootconfig.

This change allows Android S devices that launched with 4.19 upgrade
their kernels without needing to meet this new requirement.

Test: atest VtsBootconfigTest
Bug: 321122900
Change-Id: I540e38fd3f5c41d5f18011c14e3a7a2e42d57110