Press `Backspace` to enter fastboot
On RISCV QEMU, OpenSBI seems to redirect banner logging to console
input. The log contains 'f' character and frequently makes GBL drop to
fastboot. As a temporarly workaround, the CL updates the key to be
Backspace which is less likely from in/out. Longer term, we'll use EFI
GBL protocol to determine whether fastboot should be entered.

Change-Id: If8b26661b93e876fab2a6d69596e64a8bf0ee2c7