Acquires lock before handling TCP network
Fastboot USB download speed is notably slower when the Fastboot TCP task
is allowed to handle traffic without locking the fastboot shared
resource first. This is likely because our executor is cooperative and
the network polling code turns out to have non-negligible interference
on the fastboot USB task. The CL updates the TCP task to acquire locked
shared resource first to prevent interfering with the USB task that is
in the middle of a fastboot session.

Change-Id: I002301422d0a7a414854ecf3d5380091c9c263a6