Proposal for a generic GBL partition
The CL proposes to extend the concept of partition in GBL to cover both
GPT partitions and raw storage device as a partition.

A GBL defined "Vendor-defined Device Media Path" is introduced to
specify that a raw storage device is used as a partition and its
partition name.

Example use cases include emmc boot0/boot1 hardware partitions and any
non-gpt devices.

Bug: 357688291
Change-Id: I1c598c9b1bbc5c0ca6372d30788abdcc9e46c933