libefi: mark notification callbacks unsafe
The notification callback functionality has some remaining unsafe corner
cases with the global EFI hooks which have proven very difficult to
handle safely, so for now we mark this functionality unsafe and require
the caller to properly handle it.

In practice our notification callbacks are trivial and already safe, so
this is really just protection against accidental future misuse.

This also allows simplifying some of the event creation API lifetimes.

Bug: b/361404717
Change-Id: I93c518da9305e0262c66c5269331d9f0b9897992