Abstract fastboot buffer management
Introduces a BufferPool interface for abstracting the management of
fastboot download buffers. This allows caller to implement different
dynamic buffer allocation strategy. Also removes the uses of Mutex and
switches to the rust built-in RefCell for sharing buffer pool. GBL
does its own scheduling of async tasks on single thread.

Bug: 373461015

Change-Id: I65fbca824d54370b8fb3562e1c0fe15a8f798ed5