Update backup in primary header when flashing GPT
Fixes a bug in update_gpt() that the `backup` field in the primary
header is not adjusted. This causes GPT sync to always find that the
primary and secondary header are not consistent and as as result always
restores secondary GPT header.

But because we actually can't rely on backup index for finding backup
GPT, the consistency check is now updated to simply ignore
current/backup.

Change-Id: Icf98c62d072528bc6975a18edd829a5ebd9ce378