Manually invoke tests at build time
The presubmit infra used by GBL only blocks on build failure and not on
test failure. This has caused breakage serveral times. The issue is
currently under investigation in b/382503065. To limit the impact in
the mean time, we use a temporary workaround that manually invokes test
executables at build time so that any test failure will become build
failure. Doc test is not supported however. It has to be run as a
normal test

Bug: b/382503065

Change-Id: Iac2171aac40ecbad4d6cea8d1cdcc5bc31bee165