Reduce fastboot async stack use
GBL fastboot currently puts all Future on stack and executes from it,
which causes a stack use of almost 200K. This excceeds the stack size
on sorrel and ARM64 OVMF. The CL optimizes it by moving the Future to
heap. The peak stack use is now down to about 20k upon entering
fastboot and is mostly used by `DeviceTreeComponentRegistry`(18k) which
is already a known issue under investigation.

Bug: 404271887

Change-Id: I1e2d42c970d1b4ea5892ebacf29272514525c279